requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==4.2.11
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
